src/main/java/com/tw/darkhorse/controller/DemoController.java
package com.tw.darkhorse.controller;

import com.tw.darkhorse.service.DemoService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/demo")
public class DemoController {

    private DemoService demoService;

    public DemoController(final DemoService demoService) {
        this.demoService = demoService;
    }

    @GetMapping("/foo")
    public ResponseEntity readReadings() {

        return ResponseEntity.ok("abc");
    }
}
